updated Bankrupt, deleted isOwned Boolean from Square, moved owner from PpropertySquare to Square.
