+++ src/main/java/controller/BankruptController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import model.Player;
 import model.square.property.PropertySquare;
 import ui.GUIBoundary;
 
@@ -25,6 +26,17 @@ public class BankruptController {
             // boolean isBankrupt is true
             // slet bilen fra spillerpladen.
 
+
+    }
+
+    public void transferPropertyToCreditor(PlayerController playerController, PropertySquare propertySquare, Player owner) {
+        if(owner == null) {
+            PropertySquare[] currentPlayerProperties = playerController.getCurrPlayerProperties();
+
+            for(PropertySquare square: currentPlayerProperties) {
+                square.setOwner(null);
+            }
+        }
 
     }
