+++ src/main/java/model/square/property/StreetSquare.java
@@ -20,16 +20,16 @@ public class StreetSquare extends PropertySquare {
 
 
 
-        if(super.getIsOwned() && !playerController.getCurrPlayer().equals(super.getOwner())){ //pay rent.
+        if(super.getOwner() != null && !playerController.getCurrPlayer().equals(super.getOwner())){ //pay rent.
             //TODO pay rent to owner. Do not forget overwrite setCurrScenarioForPlayer
             propertyController.payRent(this, playerController);
 
 
-        } else if (super.getIsOwned() && playerController.getCurrPlayer().equals(super.getOwner())){ //owned by current player
+        } else if (super.getOwner() != null && playerController.getCurrPlayer().equals(super.getOwner())){ //owned by current player
             super.setCurrScenarioForPlayer(playerController.getCurrPlayerName() + " står på " + super.getSquareName() +
                     " som " + playerController.getCurrPlayerName() + " ejer selv.");
 
-        } else if(!super.getIsOwned()){ //buy property
+        } else if(super.getOwner() == null){ //buy property
             propertyController.buyProperty(this, playerController);
         }
     }
